Prevent mutation of the original style
An example use case is when rendering server side, the first time after a server restart, a `:focus` selector will be rendered correctly. But the second time the pseudo selector will be missing from the original style object and the rule child won't be added.

The `hasOwnProperty` is not related, but I didn't want to make another pull request.